Use cosf and sinf instead of cos and sin to avoid few conversion warnings.

+++ pffft.c
@@ -1102,8 +1102,8 @@ static void rffti1_ps(int n, float *wa, int *ifac)
       for (ii = 3; ii <= ido; ii += 2) {
         i += 2;
         fi += 1;
-        wa[i - 2] = cos(fi*argld);
-        wa[i - 1] = sin(fi*argld);
+        wa[i - 2] = cosf(fi*argld);
+        wa[i - 1] = sinf(fi*argld);
       }
       is += ido;
     }
@@ -1137,8 +1137,8 @@ void cffti1_ps(int n, float *wa, int *ifac)
       for (ii = 4; ii <= idot; ii += 2) {
         i += 2;
         fi += 1;
-        wa[i-1] = cos(fi*argld);
-        wa[i] = sin(fi*argld);
+        wa[i-1] = cosf(fi*argld);
+        wa[i] = sinf(fi*argld);
       }
       if (ip > 5) {
         wa[i1-1] = wa[i-1];
@@ -1229,8 +1229,8 @@ PFFFT_Setup *pffft_new_setup(int N, pffft_transform_t transform) {
     int j = k%SIMD_SZ;
     for (m=0; m < SIMD_SZ-1; ++m) {
       float A = -2*M_PI*(m+1)*k / N;
-      s->e[(2*(i*3 + m) + 0) * SIMD_SZ + j] = cos(A);
-      s->e[(2*(i*3 + m) + 1) * SIMD_SZ + j] = sin(A);
+      s->e[(2*(i*3 + m) + 0) * SIMD_SZ + j] = cosf(A);
+      s->e[(2*(i*3 + m) + 1) * SIMD_SZ + j] = sinf(A);
     }
   }
